--- a/test/micro/org/openjdk/bench/java/lang/Integers.java
+++ b/test/micro/org/openjdk/bench/java/lang/Integers.java
@@ -173,4 +173,11 @@ public class Integers {
             res[i] = Integer.reverseBytes(intsSmall[i]);
         }
     }
+
+    @Benchmark
+    public void reverse() {
+        for (int i = 0; i < size; i++) {
+            res[i] = Integer.reverse(intsSmall[i]);
+        }
+    }
 }