Summary of changes:
- Intrinsify scalar bit reverse APIs to emit efficient instruction sequence for X86 targets with and w/o GFNI feature.
- Handle auto-vectorization of Integer/Long.reverse bit operations.
- Backend implementation for these were added with 4th incubation of VectorAPIs.

Following are performance number for newly added JMH mocro benchmarks:-

```
No-GFNI(CLX):
=============
Baseline:
Benchmark         (size)  Mode  Cnt  Score   Error  Units
Integers.reverse     500  avgt    2  1.085          us/op
Longs.reverse        500  avgt    2  1.236          us/op
WithOpt:
Benchmark         (size)  Mode  Cnt  Score   Error  Units
Integers.reverse     500  avgt    2  0.104          us/op
Longs.reverse        500  avgt    2  0.255          us/op

With-GFNI(ICX):
===============
Baseline:
Benchmark         (size)  Mode  Cnt  Score   Error  Units
Integers.reverse     500  avgt    2  0.887          us/op
Longs.reverse        500  avgt    2  1.095          us/op

Without:
Benchmark         (size)  Mode  Cnt  Score   Error  Units
Integers.reverse     500  avgt    2  0.037          us/op
Longs.reverse        500  avgt    2  0.145          us/op
```

Kindly review and share feedback.

Best Regards,
Jatin

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8290034](https://bugs.openjdk.org/browse/JDK-8290034): Auto vectorize reverse bit operations.


### Reviewers
 * [Xiaohong Gong](https://openjdk.org/census#xgong) (@XiaohongGong - Committer)
 * [Vladimir Kozlov](https://openjdk.org/census#kvn) (@vnkozlov - **Reviewer**) ⚠️ Review applies to [53157a29](https://git.openjdk.org/jdk/pull/9535/files/53157a291de18422b34e25f4cd15444cff9b4745)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk pull/9535/head:pull/9535` \
`$ git checkout pull/9535`

Update a local copy of the PR: \
`$ git checkout pull/9535` \
`$ git pull https://git.openjdk.org/jdk pull/9535/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 9535`

View PR using the GUI difftool: \
`$ git pr show -t 9535`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/9535.diff">https://git.openjdk.org/jdk/pull/9535.diff</a>

</details>